PSO Fitness v1.0
Picit komolyabb fitness vizsgálat, de nem tökéletes.

+++ backend/cpp/pso_fingering_optimization/note_position.h
@@ -6,7 +6,7 @@ class NotePosition
 {
 public:
     NotePosition(const int newStringIdx, const int newFretIdx);
-    int Distance(const NotePosition &otherPos); // távolság 2 pozíció között
+    double Distance(const NotePosition &otherPos); // távolság 2 pozíció között
     int GetStringIdx() const;
     int GetFretIdx() const;
     std::string ToString() const;
